Add a Cargo override.
This way, when we build the filament-ext library for this project, it uses this project's mogilefs_common library rather than one from git. Rust will happily compile two different versions of the mogilefs_common library, leading to weird cases where the traits the filament-ext library implements are not the same as the traits provided in mogilefs_common, despite them having the same names / paths.